requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.9
langchain-groq==0.1.6
langchainhub==0.1.18
langchain-nvidia-ai-endpoints==0.2.0 

huggingface-hub==0.36.0
langsmith==0.1.147      
langchain-huggingface==0.0.3
pydantic==2.12.3
typing-extensions==4.15.0
annotated-types==0.7.0
pydantic-core==2.41.4
typing-inspection==0.4.2
zstandard==0.25.0

faiss-cpu==1.8.0.post1
chromadb==0.5.3
cassio==0.1.10

unstructured==0.15.3
wikipedia==1.4.0
pypdf==6.1.3
langchain-text-splitters==0.2.2
bs4==0.0.2
pyPDF2

streamlit==1.38.0
ipykernel==7.0.1
python-dotenv==1.1.1
datasets
gradio